Moved the gl module into a more general ffi module
